Move type defs and parsers from account to types/utils
